requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.3/constraints-3.11.txt"
apache-airflow-providers-amazon
boto3>=1.34.0
pandas>=1.5.0
requests>=2.31.0




